% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{ad_crit}
\alias{ad_crit}
\title{Criterion function for ad_halt.}
\usage{
ad_crit(covariate, condition)
}
\arguments{
\item{covariate}{A vector containing a covariate to match the conditions on.}

\item{condition}{A factor vector containing condition labels.}
}
\description{
Criterion function for ad_halt.
}

