% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_crit}
\alias{apply_crit}
\title{Returns smallest halting_test() / thresh ratio, or zero if less than 1.}
\usage{
apply_crit(covariates, crit, condition, thresh)
}
\arguments{
\item{covariates}{A vector or columnwise matrix containing
                     covariates to match the conditions on.}

\item{crit}{The criterion function to use, such as \code{\link{t_crit}}.}

\item{condition}{A factor vector containing condition labels.}

\item{thresh}{The statistical threshold to pass onto the
                     aforementioned test.}
}
\description{
Returns smallest halting_test() / thresh ratio, or zero if less than 1.
}

