% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search_exhaustive.R
\name{combine_ratios}
\alias{combine_ratios}
\title{Combines matched group candidate with current best one.}
\usage{
combine_ratios(best, candidate)
}
\arguments{
\item{best}{The best matched groups so far together with its
                  p-value / thresh ratio; a list containing ratio and inds
                  (a list of subject index vectors).}

\item{candidate}{A list containing ratio (a number) and ind (a vector).}
}
\value{
A structure like best containing the ind vectors with the highest
ratio.
}
\description{
Combines matched group candidate with current best one.
}

