% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{create_halting_test}
\alias{create_halting_test}
\title{Creates halt test from multiple tests.}
\usage{
create_halting_test(halting_tests)
}
\arguments{
\item{halting_tests}{A vector of halt test functions with the signature:
halting_test(condition, covariates, thresh).
For the meaning of the parameters see
\code{\link{ldamatch}}.}
}
\value{
A function that returns the minimum of all halting test values;
        the threshold value supplied to it is recycled for the individual
        functions.
}
\description{
The created halt test function returns the smallest p-value / threshold
ratio of the values produced by the supplied halt test functions.
}

