% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ldamatch.R
\name{create_subject_subspace_using_LDA}
\alias{create_subject_subspace_using_LDA}
\title{Creates ordered subspace of subject candidates for removal, using LDA.}
\usage{
create_subject_subspace_using_LDA(condition, covariates)
}
\arguments{
\item{condition}{A factor vector containing condition labels.}

\item{covariates}{A vector or columnwise matrix containing
                     covariates to match the conditions on.}
}
\value{
An ordered subject subspace: a list of vectors, with one vector per
group containing the corresponding subject indices.
}
\description{
Creates ordered subspace of subject candidates for removal, using LDA.
}

