% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{.apply_crit_to_condition_pairs}
\alias{.apply_crit_to_condition_pairs}
\title{Returns smallest value from .apply_crit for all condition pairs.}
\usage{
.apply_crit_to_condition_pairs(covariates, crit, condition, thresh)
}
\arguments{
\item{covariates}{A columnwise matrix containing
covariates to match the conditions on.}

\item{crit}{The criterion function to use, such as \code{\link{.t_crit}}.}

\item{condition}{A factor vector containing condition labels.}

\item{thresh}{The return value of halting_test has to be greater than
or equal to thresh for the matched groups.}
}
\value{
The ratio of the p-value and the threshold, or 0 if the p-value is
        less than the threshold.
}
\description{
Returns smallest value from .apply_crit for all condition pairs.
}
\keyword{internal}
