% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_random.R
\name{.calc_multipliers}
\alias{.calc_multipliers}
\title{Calculates multipliers used in search_random.}
\usage{
.calc_multipliers(counts, rcounts, props)
}
\arguments{
\item{counts}{The number of subjects for each group.}

\item{rcounts}{The number of subjects that can be removed for each group.}

\item{props}{The expected proportion of subjects for each group.}
}
\description{
Derives multiplier for rcounts (the number of subjects that can be removed)
such that the proportion of the expected sizes of groups will be props.
The returned multipliers will be in the range of 0 to 1.
}
\keyword{internal}
