% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_heuristic4.R
\name{.choose_most_frequently_chosen_subject_from_subject_tuples}
\alias{.choose_most_frequently_chosen_subject_from_subject_tuples}
\title{Chooses best one(s) of a set of subjects having the best p-value(s).}
\usage{
.choose_most_frequently_chosen_subject_from_subject_tuples(
  is.in,
  best_sets,
  look,
  condition,
  covariates,
  halting_test,
  thresh,
  tiebreaker,
  props,
  prefer_test,
  max_removed_per_cond,
  max_removed_in_next_step,
  ratio_for_slowdown,
  remove_best_only,
  print_info
)
}
\value{
The table of counts for the chosen indices within is.in.
}
\description{
Used as first parameter of .search_heuristic_with_lookahead.
It chooses subject(s) for removal: the most frequently removed one(s);
among those, it prefers the one with the lowest p-value,
then chooses randomly among the remaining ones.
}
\keyword{internal}
