% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.combine_sets}
\alias{.combine_sets}
\title{Combines current best and candidate sets, keeping the highest metric value.}
\usage{
.combine_sets(best, candidate)
}
\arguments{
\item{best}{A list(metric, sets); metric is a number, set is a list of
vectors.}

\item{candidate}{A list(metric, set); metric is a number, set is a vector.
Candidate is only considered if metric is not zero.}
}
\value{
A list containing the highest metric and a list of set values (sets).
}
\description{
Combines current best and candidate sets, keeping the highest metric value.
}
\keyword{internal}
