% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.flip_ind}
\alias{.flip_ind}
\title{Flips logical vector at specified indices}
\usage{
.flip_ind(ind, is.in)
}
\arguments{
\item{ind}{Integer indices for the is.in logical vector.}

\item{is.in}{A logical vector showing which items are preserved.}
}
\value{
A logical vector identical to is.in except for indices in ind where
        it is is.in negated.
}
\description{
Flips logical vector at specified indices
}
\keyword{internal}
