% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{.get_halting_test}
\alias{.get_halting_test}
\title{Returns halting tests for names, or checks if pass functions are suitable.}
\usage{
.get_halting_test(halting_test)
}
\arguments{
\item{halting_test}{The name of one halting test, or a halting test function.}
}
\value{
A vector of halting test functions.
}
\description{
Returns halting tests for names, or checks if pass functions are suitable.
}
\keyword{internal}
