% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldamatch.R
\name{.get_if_args_are_missing}
\alias{.get_if_args_are_missing}
\title{Determines which arguments for a function, which is its caller by default.}
\usage{
.get_if_args_are_missing(fun = sys.function(-1), ncall = 3)
}
\arguments{
\item{fun}{A function; default: the caller.}

\item{ncall}{The parent frame index; default: 3 (the great-grandparent).}
}
\value{
A named boolean vector that contains whether each argument is missing.
}
\description{
Determines which arguments for a function, which is its caller by default.
}
