% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldamatch.R
\name{.normalize_max_removed_per_cond}
\alias{.normalize_max_removed_per_cond}
\title{Normalizes max_removed_per_cond parameter for match_groups() and estimate_exhaustive().}
\usage{
.normalize_max_removed_per_cond(max_removed_per_cond, condition)
}
\arguments{
\item{max_removed_per_cond}{A named integer vector, containing the maximum number
of subjects that can be removed from each group.
Specify 0 for groups if you want to preserve
all of their subjects. If you do not specify a value
for a group, it defaults to 2 less than the group size.
Values outside the valid range of 0..(N-1)
(where N is the number of subjects in the group)
are corrected without a warning.}

\item{condition}{A factor vector containing condition labels.}
}
\description{
Normalizes max_removed_per_cond parameter for match_groups() and estimate_exhaustive().
}
\keyword{internal}
