% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{.recycle}
\alias{.recycle}
\title{Recycles threshold values for halting tests.}
\usage{
.recycle(ts, hs)
}
\arguments{
\item{ts}{Threshold value(s).}

\item{hs}{Halting tests.}
}
\value{
A vector with one threshold value per halting test.
}
\description{
Recycles threshold values for halting tests.
}
\keyword{internal}
