% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{.t_crit}
\alias{.t_crit}
\title{Criterion function for t_halt.}
\usage{
.t_crit(covariate, condition)
}
\arguments{
\item{covariate}{A vector containing a covariate to match the conditions on.}

\item{condition}{A factor vector containing condition labels.}
}
\value{
The p-value.
}
\description{
Criterion function for t_halt.
}
\keyword{internal}
