% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{.tolerance}
\alias{.tolerance}
\title{An infinitesimally small amount, used to check if values are
approximately the same.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.tolerance
}
\description{
An infinitesimally small amount, used to check if values are
approximately the same.
}
\keyword{internal}
