% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.unique_list}
\alias{.unique_list}
\title{Uniquifies a list.}
\usage{
.unique_list(l)
}
\arguments{
\item{l}{A list.}
}
\value{
The unique list items.
}
\description{
Uniquifies a list.
}
\keyword{internal}
