% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.vector_list_to_string}
\alias{.vector_list_to_string}
\title{Creates string from list of vectors.}
\usage{
.vector_list_to_string(lv, sep = "")
}
\arguments{
\item{lv}{A list of vectors.}

\item{sep}{A string to be inserted between the name of a vector item and
its value.}
}
\value{
A character string.
}
\description{
Creates string from list of vectors.
}
\keyword{internal}
