% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_exhaustive.R
\name{get_human_readable}
\alias{get_human_readable}
\title{Returns human readable format for number of seconds.}
\usage{
get_human_readable(seconds, num_decimals = 3)
}
\arguments{
\item{seconds}{The number of seconds to convert to human-readable form.}

\item{num_decimals}{The number of decimals to print in the output.}
}
\value{
A string containing "<number> seconds/minutes/hours/days/years".
}
\description{
Returns human readable format for number of seconds.
}

