% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_param}
\alias{get_param}
\title{Gets value for ldamatch global parameter.}
\usage{
get_param(name)
}
\arguments{
\item{name}{The name of the global parameter.}
}
\value{
The value of the global parameter.
}
\description{
Gets value for ldamatch global parameter.
}
\seealso{
\code{\link{set_param}} for parameter names.
}
