% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ldamatch.R
\docType{package}
\name{GUtils}
\alias{GUtils}
\alias{GUtils-package}
\title{ldamatch: Multivariate Condition Matching by Backwards Elimination using Fisher's Linear Discriminant}
\description{
Performs group matching by backward elimination using linear discriminant analysis.
}

