% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{ks_crit}
\alias{ks_crit}
\title{Criterion function for ks_halt.}
\usage{
ks_crit(covariate, logical_condition)
}
\arguments{
\item{covariate}{A vector containing a covariate to match the conditions on.}

\item{logical_condition}{A logical vector separating the groups.}
}
\description{
Criterion function for ks_halt.
}

