% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldamatch.R
\docType{package}
\name{ldamatch-package}
\alias{ldamatch}
\alias{ldamatch-package}
\title{ldamatch: Selection of Statistically Similar Research Groups.}
\description{
Select statistically similar research groups by backward selection
using various robust algorithms,
including a heuristic based on linear discriminant analysis,
multiple heuristics based on the test statistic,
and parallelized exhaustive search.
See the help for function \code{\link{match_groups}}.
}
\author{
\strong{Maintainer}: Kyle Gorman \email{kylebgorman@gmail.com}

Authors:
\itemize{
  \item Géza Kiss \email{kiss2017@alumni.ohsu.edu}
}

}
\keyword{internal}
