% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldamatch.R
\docType{data}
\name{matching_methods}
\alias{matching_methods}
\title{The available methods for matching.}
\format{
An object of class \code{character} of length 5.
}
\usage{
matching_methods
}
\description{
The available methods for matching.
}
\keyword{datasets}
