% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldamatch.R
\docType{data}
\name{nondeterministic_matching_methods}
\alias{nondeterministic_matching_methods}
\title{The available nondeterministic methods for matching.}
\format{
An object of class \code{character} of length 3.
}
\usage{
nondeterministic_matching_methods
}
\description{
The available nondeterministic methods for matching.
}
\keyword{datasets}
