% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldamatch.R
\docType{data}
\name{parallelized_matching_methods}
\alias{parallelized_matching_methods}
\title{The available parallelized methods for matching.}
\format{
An object of class \code{character} of length 3.
}
\usage{
parallelized_matching_methods
}
\description{
The available parallelized methods for matching.
}
\keyword{datasets}
