% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{t_halt}
\alias{t_halt}
\title{A univariate halting test using the t-test (t.test).}
\usage{
t_halt(condition, covariates, thresh)
}
\arguments{
\item{condition}{A factor vector containing condition labels.}

\item{covariates}{A vector or columnwise matrix containing
                     covariates to match the conditions on.}

\item{thresh}{The statistical threshold to pass onto the
                     aforementioned test.}
}
\description{
A univariate halting test using the t-test (t.test).
}

