% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{wilks_halt}
\alias{wilks_halt}
\title{A multivariate halting test appropriate for more than two condition levels.}
\usage{
wilks_halt(condition, covariates, thresh)
}
\arguments{
\item{condition}{A factor vector containing condition labels.}

\item{covariates}{A columnwise matrix containing
covariates to match the conditions on.}

\item{thresh}{The return value of halting_test has to be greater than
or equal to thresh for the matched groups.}
}
\value{
The ratio of the p-value and the threshold, or 0 if the p-value is
        less than the threshold.
}
\description{
A multivariate halting test appropriate for more than two condition levels.
}
