% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{ks_halt}
\alias{ks_halt}
\title{A univariate halting test using the Kolmogorov-Smirnov Test (ks.test).}
\usage{
ks_halt(condition, covariates, thresh)
}
\arguments{
\item{condition}{A factor vector containing condition labels.}

\item{covariates}{A vector or columnwise matrix containing
                     covariates to match the conditions on.}

\item{thresh}{The statistical threshold to pass onto the
                     aforementioned test.}
}
\description{
Note that the condition must have two levels.
The null hypothesis is that they are drawn from the same distribution,
so we want this to be *likely*. It is up to you to choose an
appropriate threshold.
}

