% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldamatch.R
\docType{package}
\name{ldamatch}
\alias{ldamatch}
\alias{ldamatch-package}
\title{ldamatch: Selection of Statistically Similar Research Groups.}
\description{
Select statistically similar research groups by backward selection
using various robust algorithms,
including a heuristic based on linear discriminant analysis,
multiple heuristics based on the test statistic,
and parallelized exhaustive search.
See the help help for function \code{\link{match_groups}}.
}

