% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/globals.R
\name{set_param}
\alias{set_param}
\title{Sets parameters for ldamatch.}
\usage{
set_param(name, value)
}
\arguments{
\item{name}{The name of the global parameter.}

\item{value}{The new value of the global parameter.}
}
\value{
The previous value of the global parameter.
}
\description{
Sets parameters for ldamatch.
}
\details{
The names of the available parameters:
\itemize{
  \item{MC_DEFAULT_REPLICATES}{: Monte Carlo search: default number of replicates}
  \item{Anderson-Darling test parameters; see kSamples::ad.test for explanation}{
    \itemize{
      \item{AD_METHOD}{: the method parameter for ad.test; default: asymptotic}
      \item{AD_NSIM}{: the Nsim parameter for ad.test; default: 10000}
      \item{AD_VERSION}{: 1 or 2 for the two versions of the test statistic; default: 1}
    }
  }
  \item{PRINT_INFO}{: print summary information, and progress information for
    the exhaustive search algorithm}
}
}

