% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/halting_tests.R
\name{wilks_halt}
\alias{wilks_halt}
\title{A multivariate halting test (appropriate when nlevels(condition) > 2)).}
\usage{
wilks_halt(condition, covariates, thresh)
}
\arguments{
\item{condition}{A factor vector containing condition labels.}

\item{covariates}{A vector or columnwise matrix containing
                     covariates to match the conditions on.}

\item{thresh}{The statistical threshold to pass onto the
                     aforementioned test.}
}
\description{
A multivariate halting test (appropriate when nlevels(condition) > 2)).
}

